se arreglo el crud prodctos y su conexion con la base de datos sql

+++ schemas/prod.py
@@ -5,4 +5,4 @@ from pydantic import BaseModel
 class Product(BaseModel):
     id: Optional[int] = None
     prodName: str
-    prodPrice: float
+    prodPrice: int
